Transpose dimensions in sfc_alb_dir and sfc_alb_dif
Transpose dimensions in sfc_alb_dir and sfc_alb_dif to be consistent
with the dimension ordering with the rest of the variables. Previously,
these had dimension (nband,ncol), while everything internal to the code
has dimension order (ncol,ngpt). In fact, in rte_sw, there was a call to
expand_and_transpose to take these inputs with dimension (nband,ncol)
and transpose them to (ncol,ngpt) while also expanding from bands to
gpts. I think this was because it was assumed that models would carry
around albedo with dimension nband,ncol. This is not the case for
SCREAM, nor for E3SM, so making this assumption just means we have an
extra transpose operation. While this was done along with expanding
bands to gpoints, so cost was probably minimal, it is an unneccesary
complication to assume a different dimension order for one field, so
instead we require the user to pass these fields with a dimension
ordering consistent with the rest of the code.